refactor(query): extract Canary.Query.Incidents read model
Move from lib/canary/query.ex into lib/canary/query/incidents.ex:
- Public: active_incidents/1
- Privates: maybe_filter_incident_annotation/2, annotated_incident_ids/2,
  active_incident_view/2, format_incident/3, signal_active_for_report?/2
  (all clauses), incident_severity/2, within_incident_window?/2
- Module attr: @incident_active_window_seconds

Facade now exposes active_incidents via defdelegate. Incident and
IncidentSignal aliases dropped from the facade.
